vehicle reorg: move vehicle_part into its own file
split off a couple of functions that should be vehicle_part functions
but aren't at the same time.
